Move platform enum to platform.param.ts
